+++ Sources/RubyGatewayHelpers/include/rbg_helpers.h
@@ -42,12 +42,12 @@ VALUE rbg_inspect_protect(VALUE value, int * _Nonnull status);
 
 /// Safely call `rb_funcallv` and report exception status.
 VALUE rbg_funcallv_protect(VALUE value, ID id,
-                           int argc, const VALUE * _Nonnull argv,
+                           int argc, const VALUE * _Nonnull argv, int kwArgs,
                            int * _Nonnull status);
 
 /// Safely call `rb_yield_values2` and report exception status.
 VALUE rbg_yield_values(int argc,
-                       const VALUE * _Nonnull argv,
+                       const VALUE * _Nonnull argv, int kwArgs,
                        int * _Nonnull status);
 
 /// Things a Swift callback can ask Ruby to do
@@ -96,7 +96,7 @@ void rbg_register_value_block_proc_callback(Rbg_value_block_call _Nonnull);
 /// block handler with the given context as the block.
 /// And report exception status.
 VALUE rbg_block_call_pvoid_protect(VALUE value, ID id,
-                                   int argc, const VALUE * _Nonnull argv,
+                                   int argc, const VALUE * _Nonnull argv, int kwArgs,
                                    void * _Nonnull context,
                                    int * _Nonnull status);
 
@@ -104,7 +104,7 @@ VALUE rbg_block_call_pvoid_protect(VALUE value, ID id,
 /// block handler with the given context as the block.
 /// And report exception status.
 VALUE rbg_block_call_value_protect(VALUE value, ID id,
-                                   int argc, const VALUE * _Nonnull argv,
+                                   int argc, const VALUE * _Nonnull argv, int kwArgs,
                                    VALUE context,
                                    int * _Nonnull status);
 
@@ -171,7 +171,7 @@ void rbg_inject_module_protect(VALUE into, VALUE module,
                                Rbg_inject_type type,
                                int * _Nonnull status);
 
-VALUE rbg_call_super_protect(int argc, const VALUE * _Nonnull argv,
+VALUE rbg_call_super_protect(int argc, const VALUE * _Nonnull argv, int kwArgs,
                              int * _Nonnull status);
 
 /// Callback into Swift code for gvar access
@@ -191,13 +191,18 @@ ID rbg_create_virtual_gvar(const char * _Nonnull name, int readonly);
 const char * _Nonnull rbg_ruby_version(void);
 const char * _Nonnull rbg_ruby_description(void);
 
-/// Cross Ruby version support
-unsigned long rbg_fix2ulong(VALUE v);
-long          rbg_fix2long(VALUE v);
-
 /// Horrible casts rejected by importer
 typedef void rbg_unblock_function_t(void * _Nullable);
 rbg_unblock_function_t * _Nonnull rbg_RUBY_UBF_IO(void);
+
+/// Ruby 3 incompatible changes from Swift's point of view
+int rbg_type(VALUE v);
+int rbg_qfalse(void);
+int rbg_qtrue(void);
+int rbg_qnil(void);
+int rbg_qundef(void);
+int rbg_RB_TEST(VALUE v);
+int rbg_RB_NIL_P(VALUE v);
 
 /// Rbg_value - keep a VALUE safe so it does not get GC'ed
 typedef struct  {